Fix/opensbli tests

+++ apps/c/lowdim_test/test.sh
@@ -1,54 +1,30 @@
 #!/bin/bash
 set -e
 cd ../../../ops/c
-<<COMMENT
-if [ -x "$(command -v enroot)" ]; then
-  cd -
-  enroot start --root --mount $OPS_INSTALL_PATH/../:/tmp/OPS --rw cuda112hip sh -c 'cd /tmp/OPS/apps/c/lowdim_test; ./test.sh'
-  grep "PASSED" perf_out
-  rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-  rm perf_out
-  echo "All HIP complied applications PASSED"
-fi
 
-if [[ -v HIP_INSTALL_PATH ]]; then
-  source ../../scripts/$SOURCE_HIP
-  make -j -B
-  cd -
-  make clean
-  rm -f .generated
-  make lowdim_hip lowdim_mpi_hip -j
-     
-  echo '============> Running HIP'
-  ./lowdim_hip OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
-  #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
-  #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
-  grep "PASSED" perf_out
-  rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-  rm -rf perf_out output.h5
-  
-  echo '============> Running MPI+HIP'
-  mpirun --allow-run-as-root -np 2 ./lowdim_mpi_hip OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
-  #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
-  #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
-  grep "PASSED" perf_out
-  rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-  rm -rf perf_out output.h5
-  echo "All HIP complied applications PASSED : Moving no to Intel Compiler Tests " > perf_out
-  exit 0
-fi
-COMMENT
+export SOURCE_INTEL=source_intel_2021.3_pythonenv
+export SOURCE_PGI=source_pgi_nvhpc_23_pythonenv
+export SOURCE_INTEL_SYCL=source_intel_2021.3_sycl_pythonenv
+export SOURCE_AMD_HIP=source_amd_rocm-5.4.3_pythonenv
 
+#export AMOS=TRUE
+#export DMOS=TRUE
+#export TELOS=TRUE
+#export KOS=TRUE
+
+if [[ -v TELOS || -v KOS ]]; then
+
+#============================ Test with Intel Classic Compilers==========================================
+echo "Testing Intel classic complier based applications ---- "
 cd $OPS_INSTALL_PATH/c
 source ../../scripts/$SOURCE_INTEL
-make -j -B
+make clean
+make
+
 cd $OPS_INSTALL_PATH/../apps/c/lowdim_test
 make clean
-rm -f .generated
-make IEEE=1 -j
+make IEEE=1
 
-
-#============================ Test lowdim with Intel Compilers ==========================================
 echo '============> Running SEQ'
 ./lowdim_dev_mpi > perf_out
 mv output.h5 output_seq.h5
@@ -88,6 +64,7 @@ grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 
+if [[ -v CUDA_INSTALL_PATH ]]; then
 echo '============> Running CUDA'
 ./lowdim_cuda OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
@@ -107,59 +84,71 @@ rm -rf perf_out output.h5
 #echo '============> Running MPI+CUDA with GPU-Direct'
 #MV2_USE_CUDA=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_cuda -gpudirect OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 
-
 #grep "PASSED" perf_out
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 #rm -rf perf_out output.h5
 
-echo '============> Running OpenCL on CPU'
-./lowdim_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=512 OPS_BLOCK_SIZE_Y=1 > perf_out
+fi
+
+echo "All Intel classic complier based applications ---- PASSED"
+
+fi
+
+if [[ -v TELOS ]]; then
+
+#============================ Test with Intel SYCL Compilers==========================================
+echo "Testing Intel SYCL complier based applications ---- "
+cd $OPS_INSTALL_PATH/c
+source ../../scripts/$SOURCE_INTEL_SYCL
+
+make clean
+make
+cd $OPS_INSTALL_PATH/../apps/c/lowdim_test
+
+make clean
+make IEEE=1 lowdim_sycl lowdim_mpi_sycl lowdim_mpi_sycl_tiled
+
+echo '============> Running SYCL on CPU'
+./lowdim_sycl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=512 OPS_BLOCK_SIZE_Y=1 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 
-echo '============> Running OpenCL on GPU'
-./lowdim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-./lowdim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
+echo '============> Running MPI+SYCL on CPU'
+$MPI_INSTALL_PATH/bin/mpirun -np 12 ./lowdim_mpi_sycl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 
-echo '============> Running MPI+OpenCL on CPU'
-$MPI_INSTALL_PATH/bin/mpirun -np 20 ./lowdim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
-$MPI_INSTALL_PATH/bin/mpirun -np 20 ./lowdim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
+echo '============> Running MPI+SYCL+Tiled on CPU'
+$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_sycl_tiled OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 
-echo '============> Running MPI+OpenCL on GPU'
-$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-#$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
-#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm -rf perf_out output.h5
+echo "All Intel SYCL complier based applications ---- PASSED"
 
-echo "All Intel complied applications PASSED : Moving no to PGI Compiler Tests "
+fi
 
+if [[ -v TELOS ]]; then
 
-cd -
+#============================ Test with PGI Compilers==========================================
+echo "Testing PGI/NVHPC complier based applications ---- "
+cd $OPS_INSTALL_PATH/c
 source ../../scripts/$SOURCE_PGI
-
 make clean
-make -j
-cd -
+#make -j
+make
+cd $OPS_INSTALL_PATH/../apps/c/lowdim_test
 make clean
-make -j
+make IEEE=1
 
-#============================ Test lowdim with PGI Compilers ==========================================
 echo '============> Running OpenMP'
 KMP_AFFINITY=compact OMP_NUM_THREADS=12 ./lowdim_openmp > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
@@ -192,6 +181,7 @@ grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 
+if [[ -v CUDA_INSTALL_PATH ]]; then
 echo '============> Running CUDA'
 ./lowdim_cuda OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
@@ -223,44 +213,18 @@ rm -rf perf_out output.h5
 #grep "PASSED" perf_out
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 #rm -rf perf_out output.h5
+fi
 
-echo '============> Running OpenCL on GPU'
-./lowdim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-./lowdim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-#$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
-#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm -rf perf_out output.h5
-
-#echo '============> Running MPI+OpenCL on CPU'
-#$MPI_INSTALL_PATH/bin/mpirun -np 20 ./lowdim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
-#$MPI_INSTALL_PATH/bin/mpirun -np 20 ./lowdim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
-
-
-#grep "PASSED" perf_out
-#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-#rm -rf perf_out output.h5
-
-echo '============> Running MPI+OpenCL on GPU'
-$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-#$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
-#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm -rf perf_out output.h5
-
-echo '============> Running OpenACC'
-./lowdim_openacc OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
+echo '============> Running OMPOFFLOAD'
+./lowdim_ompoffload OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 #COMMENT
-echo '============> Running MPI+OpenACC'
-$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_openacc numawrap2 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
+echo '============> Running MPI+OMPOFFLOAD'
+$MPI_INSTALL_PATH/bin/mpirun -np 2 ./lowdim_mpi_ompoffload numawrap2 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
 #$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
 grep "PASSED" perf_out
@@ -268,3 +232,40 @@ rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm -rf perf_out output.h5
 echo "All PGI complied applications PASSED : Exiting Test Script "
 
+echo "All PGI complier based applications ---- PASSED"
+
+fi
+
+if [[ -v AMOS ]]; then
+
+echo "Testing AMD HIP complier based applications ---- "
+cd $OPS_INSTALL_PATH/c
+source ../../scripts/$SOURCE_AMD_HIP
+make clean
+make
+cd $OPS_INSTALL_PATH/../apps/c/lowdim_test
+
+make clean
+make lowdim_hip lowdim_mpi_hip
+
+echo '============> Running HIP'
+./lowdim_hip OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
+#$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
+#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
+grep "PASSED" perf_out
+rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
+rm -rf perf_out output.h5
+
+echo '============> Running MPI+HIP'
+mpirun --allow-run-as-root -np 2 ./lowdim_mpi_hip OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
+#$HDF5_INSTALL_PATH/bin/h5diff output.h5 output_seq.h5
+#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi;
+grep "PASSED" perf_out
+rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
+rm -rf perf_out output.h5
+
+echo "All AMD HIP complier based applications ---- PASSED"
+
+fi
+
+echo "---------- Exiting Test Script "
